requirements.txt
kaleido==0.1.0.post1
numpy
pandas
plotly
pywin32
gitpython
pillow
requests
lxml
pyautogui